+++ Pal/src/host/Linux-SGX/enclave_ocalls.h
@@ -116,6 +116,6 @@ int ocall_eventfd(unsigned int initval, int flags);
 int ocall_get_quote(const sgx_spid_t* spid, bool linkable, const sgx_report_t* report,
                     const sgx_quote_nonce_t* nonce, char** quote, size_t* quote_len);
 
-int ocall_trim_epc_pages(void* addr, unsigned int nr_pages);
+int ocall_trim_epc_pages(void* addr, size_t nr_pages);
 
-int ocall_notify_accept(void* addr, unsigned int nr_pages);
+int ocall_notify_accept(void* addr, size_t nr_pages);
